added login data

+++ DIAMOND-MERCHANT/BACKEND/DiamondMerchant/src/main/java/com/repository/CustomerRepository.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.bean.Customer;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer,Integer> {
+public interface CustomerRepository extends JpaRepository<Customer,Long> {
 
+	
 }
